Changed interface to generate()
generate() now takes 3 integer arguments: the number of repeated rows, the number of
repeated bricks in a row, the width of the largest brick. It returns a vector with all
possible walls.

Fixed is_brickwork() to ensure both rows check the minimum distance before declaring it
good. Added unit test to check where this makes a difference.

Arguments to main are now the same as for generate(). Added tests for the expected number
of walls from generate().