Draw walls with the first course at the bottom.
Un-transformed SVG coordinates have y increasing downward and stream output is typically
shown with new lines at the bottom. This makes rendered walls upside down if drawn in the
most straightforward way. Fiddle the indices when drawing so the first row is at the
bottom.

Increment widths so the leftmost brick in a pattern is the least significant (slowest
changing) width. This makes for a more orderly arrangement of rendered walls.

Added fast algorithm for counting the number of possible brickworks in the case of 2 rows
and 2 bricks. Will generalize in the future.

Parse command line options with GNU Getpot.

Add incomplete math discussion to README.